Working on sub import from mp4
git-svn-id: https://subtitleedit.googlecode.com/svn/trunk@674 99eadd0c-20b8-1223-b5c4-2a2b2df33de2

+++ src/Logic/Mp4/Boxes/Mdhd.cs
@@ -13,7 +13,7 @@ namespace Nikse.SubtitleEdit.Logic.Mp4.Boxes
         public readonly string Iso639ThreeLetterCode;
         public readonly int Quality;
 
-        public Mdhd(FileStream fs, uint size)
+        public Mdhd(FileStream fs, ulong size)
         {
             byte[] b = new byte[size - 4];
             fs.Read(b, 0, b.Length);
